+++ src/ast/literal/long-literal.ts
@@ -16,7 +16,7 @@ export class LongLiteral extends Literal {
   }
 
   public getValueInternal(_state: ExpressionState): TypedValue {
-    return new TypedValue(this.value);
+    return new TypedValue(this.value, null, 'long');
   }
 
   public toStringAST(): string {